Multiline Aware Syntax Highlighting
The scope is no longer limited to the current line

Fixes
- else and except coloring
- multiline strings

Will also allow the current tree to be used for highlighting (to get the incremental parsing optimizations). Not implemented yet because the highlight library will need to be modified to allow that.

Also reworks the text layout system so native and web are able to share a lot more code. Will make implementing future features easier because the surface area of the shim is much smaller.