Preventing Unclosed Strings
Placing the cursor after a quote of a string "pseudo-selects" the string. Hitting delete the entire string along with the quote.

Uses the syntax tree to find the starts and end of the string

This is the first test of adding more frame-based editor interactions

this currently slightly overlaps with the insertion pair deletion. if this is later expanded to more paired characters, that can be removed.