Back with Text Rope
We do a lot of operations on Unicode characters (insertion, deletion, finding the nth line, etc.). A text rope is much more efficient at those types of operations. We can also afford the minimum allocation size of a KB because we are only allocating a single rope per file.

In my completely unscientific tests, it made the VSCode extension feel a bit snappier. It is felt less for the native extension with test files (because it is already fast there) but in theory would remove a bottleneck for larger files.