Configurable Font Family and Size
Switches from constants for font properties to a once lock. The family and size are set during startup. When the editor appears for the first time, the dimensions of the font are calculated (because that is the first point where the piet text factory is availabile).

The properties are then unwrapped on every read because something needing to know the size before before the widget is added to the tree would be architecturally incorrect.

VSCode passes its font info to the editor by having a separate script tag that is read by the actual one. This is a hacky solution and should be changed once Lilypad is packaged into a proper library to be integrated with other editors.