Update Python Grammar & Fix Overeager Pseudo Selections
On the grammar side, upstream now has a C scanner so my wasm compatilibty stuff was updated to use that instead of my ported rust scanner.

The new grammar version changes the way that strings are handled. Pseudo selections have been adjusted to account for this. This also fixes the bugs where quotes in comments or as a part of escape sequences would cause incorrect pseudo selections.