Fix Rare Text Rope Crash
Just using as_str on a RopeSlice fails if the slice crosses a chunk boundary (rarely happens for strings the size of lines but is still possible). Using Cow::from will automatically fallback if as_str is None.

This fixes text file 2 crashing on wasm, and rendering a line incorrectly on native.