Fix crash when rapidly deleting and selecting at end of file
Padding is used to calculate what is being selected. It is updated at draw time which can cause it to get out of sync with the actual text when rapidly deleting and selecting. This catches that and just clamps the line count.

I don't think it's worth catching the root cause of it getting out of sync because this sort of rapid input is more of something to survice than something to perfectly represent.