Code Completion
Ties into the VSCode code completion system. Shows the completions using a popup similar to the diagnostic popup.

Currently very overeager in what it shows and when. Actual VSCode hides a lot of the suggestions. Doing something closer to that would make this more useful.

also changes (to make the completion experience better):
- Diagnostics now pop up on hover instead of edit. Having both the diagnostics and the code completions showing at once when editing was disorienting because it hid both what was above and below.
- error block outlines are hidden. Error blocks appearing and dissapearing was messing with padding which was extra disorienting when reading the completion popup. Feature flags it so it it needs to be reinabled it can be.