Architecture for swapping languages & basic java support
Abstracts language functionality into a LanguageConfig struct so new languages can be added easier in the future.

Also adds Java support. This was done to verify that the abstraction does not have python only assumptions baked into it. Java support is fully functioning, however the editing experience is still not as nice as Python's. For example:
- Hitting newline inside a {} does not also move the closing bracket to a third line
- Code completions are overeager (the heuristics for not showing on python didn't all translate well to how the java completions work)

Also adds better adherence to the VSCode completion and diagnostics specs because assumptions that were true for Python are no longer true for Java