Bandaid For Wrong Web Height
For some reason, when the view first loads, druid sets the height of the canvas element to the wrong value. This sends a resize event after loading so it will recalculate and be set to the correct value after load.

This causes some jank where it resizes after appearing. It would be better if it just had the correct height from the start but this will do for now.