Dragging Blocks
Holding alt allows for clicking on blocks to pick them up and drop them elsewhere.

It also has the text that is dragged appear in a popup by the mouse so you actually look like you're dragging something.

This implementation addresses one challenge with drag and drop in Python. Since there aren't closing scope characters, the indentation of what will be dropped cannot be assumed. To address this, the indentation is selected by the row that you drop it into (reflected by the start of the horizontal line cursor) and the indentation of the dropped block is adjusted to reflect that selection. This is currently also enabled for {} languages but should replaced with just assuming the indentation there.

A potential improvement for this would be to add a spacer where the block was lifted from equal to the height of the block (that lasts until drop), so that things don't shift around when intially picking up a block.

This dragging system could be shared by the block palette when that is added.

This also fixes a couple bugs I ran into when implementing this:
- Pasting something that ended in a newline would place the cursor in an unexpected place. It now will be after that newline.
- Unindenting when less than 4 would crash
- Indenting when indent not divisible by 4 could set the cursor to an invalid position
- Lifting the mouse outside the window would keep selecting text when entering back
- Copy, paste, and VSCode edit application would feel really slow because it was not triggering a redraw

Also does some small refactoring
- Generalized the block and text drawing
- Moves commands into their own module so they don't need to all end in `SELECTOR`