2nd Major Text Editing Refactor
This time it is centered around making it much more testable.

Text edits now run independent of the state of the widget and produce TextEdits. In the actual application, the TextEdit is applied to the widget buffer and the tree. In the tests, it can be applied to a test rope.

(un)indentation was also reworked to both fix some selection bugs and to only send a single update to VSCode so it no longer undoes each line separately.

This also includes some helpers to make writing tests easier. Notably a rope and selection can be generated using a string with `→` denoting the start of the selection and `←` denoting the end.