Fix crash when applying edit that decreases line count
Applying an edit was using the old selection to find a new pseudo selection after applying the edit, so if the edit decreased the number of lines to no longer include the selection, it would crash. Add doc comment to note this for the future.

Also fix insert string not modifying the selection. That was just missing. Whoops.

Also rename test_insert to test_char_insert to make it clearer that there will be a test_string_insert test in the future.