Handle Undo/Redo Natively
Previously it was handled by VSCode. Now it is handled within Lilypad.

Benefits:
- Custom behavior like undoing both the pick up and drop of a block in a single command
- Undo now groups characters into words as they are typed so undoing is much faster because it goes a word at a time instead of a character at a time
- Same behavior on all platforms. I want to shoot to avoid relying on VSCode for anything more than being a glorified LSP client and file picker.

Ties it into the existing undo/redo vscode command. This is a little more hacky than I would prefer but I’m fairly confident this is the best way because VSCode does not provide a clear way to do this.

Being able to rapidly apply large edits revealed a race condition in the extension. Edits were applied async and cannot be applied at the same time. There is now an explicit edit queue for the case that there are many slow to apply edits.