Update to Tree Sitter 22
Temporarily moved to a fork of tree-sitter-c2rust with some cfgs fixed. That can go away when rust FINALLY merges the flag for abi compatibility.

Reworks c shim so it actually frees the correct amount of memory. The updated python scanner started hitting issue with it because it started freeing arrays. It now stores the layout of memory allocation at the start of all allocations.