Switch from druid to egui
This is a mega commit that converts the entire interface to use egui instead of druid because it is actively maintained and should allow us to iterate more quickly due to its more flexible architecture.

This has approximate feature parity. Though there are a couple more things that are needed soon:
- Copy and paste in webassembly
- Passthrough of keypresses to vscode (for saving and running files)
- Don’t create new highlight configurations nearly as often (pass them around like the new way fonts work here in this currently)
- Switch padding to a cumulative vector for better drawing performance
- A pile of TODOs for nicities and performance improvements