Hover Documentation, Rust Support, Diagnostic Fix (#11)
Hover documentation was added. When the mouse is over a word with a valid hover object (this would be provided by another vscode extension like pylance) it now pops up. Also, rust support is now added. Diagnostics have a source string. This string is sometimes not provided by the vscode api and it is now an option.